Event enhancements

+++ db_Workflow/dbo/Tables/EventStatuses.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[EventStatuses]
 (
-	[eventStatusID] TINYINT NOT NULL,
+	[eventStatusID] SMALLINT NOT NULL,
 	[eventStatus] VARCHAR(10) NOT NULL,
 	[isTerminal] BIT CONSTRAINT [DF_EventStatuses_Terminal] DEFAULT ((0)) NOT NULL,
 	CONSTRAINT [PK_EventStatuses] PRIMARY KEY CLUSTERED ([eventStatusID] ASC),
